LayoutTests/imported/w3c/web-platform-tests/streams/piping/close-propagation-forward.sharedworker-expected.txt

CONSOLE MESSAGE: line 10: ReferenceError: Can't find variable: SharedWorker

FAIL close-propagation-forward.js shared worker wrapper file ReferenceError: Can't find variable: SharedWorker
